+++ src/main/java/com/medicalcare/medicalcareappointments/business/impl/appointment/DeleteAppointmentUseCaseImpl.java
@@ -1,6 +1,7 @@
 package com.medicalcare.medicalcareappointments.business.impl.appointment;
 
 import com.medicalcare.medicalcareappointments.business.DeleteAppointmentUseCase;
+import com.medicalcare.medicalcareappointments.exception.NotFoundAppointmentException;
 import com.medicalcare.medicalcareappointments.persistence.AppointmentRepository;
 import lombok.AllArgsConstructor;
 import org.springframework.stereotype.Service;
@@ -11,6 +12,9 @@ public class DeleteAppointmentUseCaseImpl implements DeleteAppointmentUseCase{
     private final AppointmentRepository appointmentRepository;
     @Override
     public void deleteAppointment(long id) {
+        if(appointmentRepository.findById(id).isEmpty()) {
+            throw new NotFoundAppointmentException("NOT_FOUND_APPOINTMENT");
+        }
         appointmentRepository.deleteById(id);
     }
 }
