+++ src/main/java/com/medicalcare/medicalcareappointments/business/impl/review/DeleteReviewUseCaseImpl.java
@@ -1,6 +1,7 @@
 package com.medicalcare.medicalcareappointments.business.impl.review;
 
 import com.medicalcare.medicalcareappointments.business.DeleteReviewUseCase;
+import com.medicalcare.medicalcareappointments.exception.NotFoundReviewException;
 import com.medicalcare.medicalcareappointments.persistence.ReviewRepository;
 import lombok.AllArgsConstructor;
 import org.springframework.stereotype.Service;
@@ -13,6 +14,9 @@ public class DeleteReviewUseCaseImpl implements DeleteReviewUseCase {
 
     @Override
     public void deleteReview(long id) {
+        if(reviewRepository.findById(id).isEmpty()) {
+            throw new NotFoundReviewException("NOT_FOUND_REVIEW");
+        }
         reviewRepository.deleteById(id);
     }
 }
